fix project structure, fix docker

+++ mongodb/init.sh
@@ -4,12 +4,17 @@ if test -z "$MONGODB_PASSWORD"; then
     exit 1
 fi
 
-auth="-u user -p $MONGODB_PASSWORD"
+if test -z "$MONGODB_USER"; then
+    echo "MONGODB_USER not defined"
+    exit 1
+fi
+
+auth="-u $MONGODB_USER -p $MONGODB_PASSWORD"
 
 # MONGODB USER CREATION
 (
 echo "setup mongodb auth"
-create_user="if (!db.getUser('user')) { db.createUser({ user: 'user', pwd: '$MONGODB_PASSWORD', roles: [ {role:'readWrite', db:'piggymetrics'} ]}) }"
+create_user="if (!db.getUser('user')) { db.createUser({ user: '$MONGODB_USER', pwd: '$MONGODB_PASSWORD', roles: [ {role:'readWrite', db:'piggymetrics'} ]}) }"
 until mongo piggymetrics --eval "$create_user" || mongo piggymetrics $auth --eval "$create_user"; do sleep 5; done
 killall mongod
 sleep 1
